feat(client): settings redesign -- 3 groups, sectioned card list
