--- apps/client/test/widgets/settings/accessibility_section_test.dart
@@ -1,62 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:flutter_test/flutter_test.dart';
-import 'package:shared_preferences/shared_preferences.dart';
-
-import 'package:echo_app/src/screens/settings/accessibility_section.dart';
-import 'package:echo_app/src/theme/echo_theme.dart';
-
-void main() {
-  setUp(() {
-    SharedPreferences.setMockInitialValues({});
-  });
-
-  Widget buildSection() {
-    return ProviderScope(
-      child: MaterialApp(
-        theme: EchoTheme.darkTheme,
-        darkTheme: EchoTheme.darkTheme,
-        themeMode: ThemeMode.dark,
-        home: const Scaffold(body: AccessibilitySection()),
-      ),
-    );
-  }
-
-  group('AccessibilitySection', () {
-    testWidgets('renders section title', (tester) async {
-      await tester.pumpWidget(buildSection());
-      await tester.pumpAndSettle();
-
-      expect(find.text('Accessibility'), findsOneWidget);
-    });
-
-    testWidgets('renders font size slider', (tester) async {
-      await tester.pumpWidget(buildSection());
-      await tester.pumpAndSettle();
-
-      expect(find.byType(Slider), findsOneWidget);
-    });
-
-    testWidgets('renders reduce motion toggle', (tester) async {
-      await tester.pumpWidget(buildSection());
-      await tester.pumpAndSettle();
-
-      expect(find.text('Reduce Motion'), findsOneWidget);
-    });
-
-    testWidgets('renders high contrast toggle', (tester) async {
-      await tester.pumpWidget(buildSection());
-      await tester.pumpAndSettle();
-
-      expect(find.text('High Contrast'), findsOneWidget);
-    });
-
-    testWidgets('renders font size labels', (tester) async {
-      await tester.pumpWidget(buildSection());
-      await tester.pumpAndSettle();
-
-      expect(find.text('80%'), findsOneWidget);
-      expect(find.text('150%'), findsOneWidget);
-    });
-  });
-}
