+++ apps/client/test/widgets/settings_screen_test.dart
@@ -7,7 +7,7 @@ import 'package:echo_app/src/theme/echo_theme.dart';
 
 import '../helpers/mock_providers.dart';
 
-Widget _navListApp({
+Widget _rootApp({
   SettingsSection? selected,
   void Function(SettingsSection)? onTap,
   VoidCallback? onLogout,
@@ -19,7 +19,7 @@ Widget _navListApp({
       darkTheme: EchoTheme.darkTheme,
       themeMode: ThemeMode.dark,
       home: Scaffold(
-        body: SettingsNavList(
+        body: SettingsRootView(
           selected: selected,
           onTap: onTap ?? (_) {},
           onLogout: onLogout ?? () {},
@@ -32,128 +32,117 @@ Widget _navListApp({
 void main() {
   group('settingsSectionLabel', () {
     test('returns correct label for each section', () {
-      expect(settingsSectionLabel(SettingsSection.account), 'Account');
-      expect(settingsSectionLabel(SettingsSection.devices), 'My Devices');
-      expect(settingsSectionLabel(SettingsSection.privacy), 'Privacy & Safety');
+      expect(settingsSectionLabel(SettingsSection.profile), 'Profile');
+      expect(
+        settingsSectionLabel(SettingsSection.encryptionKeys),
+        'Encryption keys',
+      );
+      expect(settingsSectionLabel(SettingsSection.devices), 'Devices');
+      expect(settingsSectionLabel(SettingsSection.appearance), 'Appearance');
       expect(
         settingsSectionLabel(SettingsSection.notifications),
         'Notifications',
       );
-      expect(settingsSectionLabel(SettingsSection.audio), 'Voice & Video');
-      expect(settingsSectionLabel(SettingsSection.appearance), 'Appearance');
-      expect(
-        settingsSectionLabel(SettingsSection.accessibility),
-        'Accessibility',
-      );
+      expect(settingsSectionLabel(SettingsSection.privacy), 'Privacy');
+      expect(settingsSectionLabel(SettingsSection.dataStorage), 'Storage');
       expect(settingsSectionLabel(SettingsSection.about), 'About');
-      expect(
-        settingsSectionLabel(SettingsSection.dataStorage),
-        'Data & Storage',
-      );
-      expect(settingsSectionLabel(SettingsSection.debug), 'Debug Logs');
     });
   });
 
   group('SettingsSection enum', () {
-    test('contains all expected sections', () {
-      expect(SettingsSection.values, hasLength(10));
+    test('contains expected 8 sections', () {
+      expect(SettingsSection.values, hasLength(8));
       expect(
         SettingsSection.values,
         containsAll([
-          SettingsSection.account,
+          SettingsSection.profile,
+          SettingsSection.encryptionKeys,
           SettingsSection.devices,
-          SettingsSection.privacy,
-          SettingsSection.notifications,
-          SettingsSection.audio,
           SettingsSection.appearance,
-          SettingsSection.accessibility,
-          SettingsSection.about,
+          SettingsSection.notifications,
+          SettingsSection.privacy,
           SettingsSection.dataStorage,
-          SettingsSection.debug,
+          SettingsSection.about,
         ]),
       );
     });
   });
 
-  group('SettingsNavList', () {
-    // The nav list needs ~700px of height to avoid overflow with all sections
-    // plus the Log Out item at the bottom. Set a tall viewport for all tests.
+  group('SettingsRootView', () {
     setUp(() {
       TestWidgetsFlutterBinding.ensureInitialized();
     });
 
-    testWidgets('renders all navigation items', (tester) async {
-      tester.view.physicalSize = const Size(800, 1200);
+    testWidgets('renders all 8 navigation rows', (tester) async {
+      tester.view.physicalSize = const Size(800, 1400);
       tester.view.devicePixelRatio = 1.0;
       addTearDown(tester.view.resetPhysicalSize);
       addTearDown(tester.view.resetDevicePixelRatio);
 
-      await tester.pumpWidget(_navListApp(selected: SettingsSection.account));
+      await tester.pumpWidget(_rootApp(selected: SettingsSection.profile));
       await tester.pumpAndSettle();
 
-      expect(find.text('Account'), findsOneWidget);
-      expect(find.text('My Devices'), findsOneWidget);
-      expect(find.text('Privacy & Safety'), findsOneWidget);
-      expect(find.text('Notifications'), findsOneWidget);
-      expect(find.text('Voice & Video'), findsOneWidget);
+      expect(find.text('Profile'), findsOneWidget);
+      expect(find.text('Encryption keys'), findsOneWidget);
+      expect(find.text('Devices'), findsOneWidget);
       expect(find.text('Appearance'), findsOneWidget);
-      expect(find.text('Accessibility'), findsOneWidget);
+      expect(find.text('Notifications'), findsOneWidget);
+      expect(find.text('Privacy'), findsOneWidget);
+      expect(find.text('Storage'), findsOneWidget);
       expect(find.text('About'), findsOneWidget);
-      expect(find.text('Data & Storage'), findsOneWidget);
-      expect(find.text('Debug Logs'), findsOneWidget);
     });
 
-    testWidgets('renders category headers', (tester) async {
-      tester.view.physicalSize = const Size(800, 1200);
+    testWidgets('renders the 3 group headers', (tester) async {
+      tester.view.physicalSize = const Size(800, 1400);
       tester.view.devicePixelRatio = 1.0;
       addTearDown(tester.view.resetPhysicalSize);
       addTearDown(tester.view.resetDevicePixelRatio);
 
-      await tester.pumpWidget(_navListApp());
+      await tester.pumpWidget(_rootApp());
       await tester.pumpAndSettle();
 
-      expect(find.text('USER SETTINGS'), findsOneWidget);
-      expect(find.text('APP SETTINGS'), findsOneWidget);
-      expect(find.text('ADVANCED'), findsOneWidget);
+      expect(find.text('ACCOUNT'), findsOneWidget);
+      expect(find.text('PREFERENCES'), findsOneWidget);
+      expect(find.text('ECHO'), findsOneWidget);
     });
 
-    testWidgets('renders Log Out button', (tester) async {
-      tester.view.physicalSize = const Size(800, 1200);
+    testWidgets('renders Log out button', (tester) async {
+      tester.view.physicalSize = const Size(800, 1400);
       tester.view.devicePixelRatio = 1.0;
       addTearDown(tester.view.resetPhysicalSize);
       addTearDown(tester.view.resetDevicePixelRatio);
 
-      await tester.pumpWidget(_navListApp());
+      await tester.pumpWidget(_rootApp());
       await tester.pumpAndSettle();
 
-      expect(find.text('Log Out'), findsOneWidget);
+      expect(find.text('Log out'), findsOneWidget);
     });
 
     testWidgets('tapping a section calls onTap', (tester) async {
-      tester.view.physicalSize = const Size(800, 1200);
+      tester.view.physicalSize = const Size(800, 1400);
       tester.view.devicePixelRatio = 1.0;
       addTearDown(tester.view.resetPhysicalSize);
       addTearDown(tester.view.resetDevicePixelRatio);
 
       SettingsSection? tapped;
-      await tester.pumpWidget(_navListApp(onTap: (s) => tapped = s));
+      await tester.pumpWidget(_rootApp(onTap: (s) => tapped = s));
       await tester.pumpAndSettle();
 
-      await tester.tap(find.text('Privacy & Safety'));
+      await tester.tap(find.text('Privacy'));
       expect(tapped, SettingsSection.privacy);
     });
 
-    testWidgets('tapping Log Out calls onLogout', (tester) async {
-      tester.view.physicalSize = const Size(800, 1200);
+    testWidgets('tapping Log out calls onLogout', (tester) async {
+      tester.view.physicalSize = const Size(800, 1400);
       tester.view.devicePixelRatio = 1.0;
       addTearDown(tester.view.resetPhysicalSize);
       addTearDown(tester.view.resetDevicePixelRatio);
 
       var loggedOut = false;
-      await tester.pumpWidget(_navListApp(onLogout: () => loggedOut = true));
+      await tester.pumpWidget(_rootApp(onLogout: () => loggedOut = true));
       await tester.pumpAndSettle();
 
-      await tester.tap(find.text('Log Out'));
+      await tester.tap(find.text('Log out'));
       expect(loggedOut, isTrue);
     });
   });
